+++ Game/game_item.cpp
@@ -301,7 +301,7 @@ int32_t game_item_modifier::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised(this->_initialised_state,
             "game_item_modifier::get_error");
-    return (game_item_modifier::_last_error);
+    return (static_cast<int32_t>(game_item_modifier::_last_error));
 }
 
 const char *game_item_modifier::get_error_str() const noexcept
@@ -1643,7 +1643,7 @@ int32_t game_item::get_error() const noexcept
     if (this->_initialised_state == FT_CLASS_STATE_UNINITIALISED)
         errno_abort_if_uninitialised(this->_initialised_state,
             "game_item::get_error");
-    return (game_item::_last_error);
+    return (static_cast<int32_t>(game_item::_last_error));
 }
 
 const char *game_item::get_error_str() const noexcept
